Reduce duplicate code in streaming service APIs
