+++ src/modules/history/streaming-services/nrk/NrkApi.ts
@@ -1,6 +1,4 @@
 import * as moment from 'moment';
-import { TraktSearch } from '../../../../api/TraktSearch';
-import { TraktSync } from '../../../../api/TraktSync';
 import { Item } from '../../../../models/Item';
 import { Errors } from '../../../../services/Errors';
 import { EventDispatcher, Events } from '../../../../services/Events';
@@ -34,13 +32,15 @@ export interface NrkProgramInfo {
 	productionYear: number;
 }
 
-class _NrkApi implements Api {
+class _NrkApi extends Api {
 	HOST_URL: string;
 	HISTORY_API_URL: string;
 	AUTH_URL: string;
 	isActivated: boolean;
 
 	constructor() {
+		super('nrk');
+
 		this.HOST_URL = 'https://tv.nrk.no';
 		this.HISTORY_API_URL = `${this.HOST_URL}/history`;
 		this.AUTH_URL = `${this.HOST_URL}/auth/token`;
@@ -126,32 +126,6 @@ class _NrkApi implements Api {
 			item = new Item({ id, type, title, year, percentageWatched, watchedAt });
 		}
 		return item;
-	};
-
-	loadTraktHistory = async () => {
-		try {
-			let promises = [];
-			const items = getStore('nrk').data.items;
-			promises = items.map(this.loadTraktItemHistory);
-			await Promise.all(promises);
-			void getStore('nrk').update();
-		} catch (err) {
-			Errors.error('Failed to load Trakt history.', err);
-			await EventDispatcher.dispatch(Events.TRAKT_HISTORY_LOAD_ERROR, { error: err as Error });
-		}
-	};
-
-	loadTraktItemHistory = async (item: Item) => {
-		if (!item.trakt) {
-			try {
-				item.trakt = await TraktSearch.find(item);
-				await TraktSync.loadHistory(item);
-			} catch (err) {
-				item.trakt = {
-					notFound: true,
-				};
-			}
-		}
 	};
 
 	convertAspNetJSONDateToDateObject = (value: string): Date => {
